Accept credentials as impl Read
Currently, the credentials are expected to be passed as a path to a file
on disk, but as type &str. These changes update the interface to accept
anything that impl Read, which means you can pass a file handle, &[u8],
buffers, etc.

This constitutes a breaking change in the interface; the burden is put
on the caller to open the file, read the file, get the credentials from
the environment, etc. However, the advantage is that the interface is
more flexible.

I also updated all the existing tests to succeed, and added a test
verifying that it also succeeds when passing the credentials as a byte
slice.

If the breaking change is a problem, we could also leave the impl Read
interface on the TokenManager itself, but provide an additional
top-level function called something like
create_shared_token_manager_from_slice, or ..._from_reader, etc. I
personally prefer the cleaner single flexible entrypoint, but it's not
my project!